+++ src/index.ts
@@ -8,9 +8,9 @@ export { generateReviewWithFallback, providerCacheIdentity, type ConfiguredRevie
 export { renderMarkdownReport, writeReports } from "./report.js";
 export { executeReview, verifyFindings, type ReviewExecution, type ReviewOptions } from "./reviewer.js";
 export { loadIssuePilotConfig } from "./issuepilot/config.js";
-export { generateIssueDescription, generateIssuePilotPlan, GeminiIssuePlanClient, OpenRouterIssuePlanClient } from "./issuepilot/generator.js";
+export { generateIssueDescription, GeminiIssueDescriptionClient, OpenRouterIssueDescriptionClient } from "./issuepilot/generator.js";
 export { GitHubIssuePilotRepository } from "./issuepilot/github.js";
-export { loadIssuePilotPlan, validateIssuePilotPlan, writeIssuePilotPlan } from "./issuepilot/plan.js";
+export { loadIssuePilotPlan, validateIssuePilotPlan } from "./issuepilot/plan.js";
 export { evaluateTaskProgress, syncIssuePilot } from "./issuepilot/sync.js";
 export type * from "./issuepilot/types.js";
 export { redactSecrets, isSensitivePath } from "./security.js";